move to pkg/config/xds
