+++ pkg/config/xds/filters.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package constants
+package xds
 
 import (
 	resource "github.com/envoyproxy/go-control-plane/pkg/resource/v3"
